pyproject.toml
[project]
name = "arduino-serial-controller"
version = "2025.05.25"
description = "Arduino Serial Controller for Unraid"
authors = [
    {name = "phred7"}
]
readme = "README.md"
requires-python = ">=3.12.10"
dependencies = [
    "pyserial>=3.5",
]

[tool.uv]
dev-dependencies = [
    "pyinstaller>=5.0",
]